Start scrapping models/ folder (#1196)
* Move split_data into end_model/data.py

* Remove unused utilities

* Merge and relocate set_seed()
