+++ snorkel/analysis/utils.py
@@ -1,3 +1,4 @@
+import random
 from typing import Dict, List
 
 import numpy as np
@@ -5,6 +6,12 @@ import scipy.sparse as sparse
 import torch
 
 from snorkel.types import ArrayLike
+
+
+def set_seed(seed: int) -> None:
+    random.seed(seed)
+    np.random.seed(seed)
+    torch.manual_seed(seed)
 
 
 def probs_to_preds(probs: np.ndarray) -> np.ndarray:
