+++ test/model/test_utils.py
@@ -5,7 +5,8 @@ import numpy as np
 import scipy.sparse as sparse
 import torch
 
-from snorkel.model.utils import recursive_merge_dicts, split_data
+from snorkel.end_model.data import split_data
+from snorkel.model.utils import recursive_merge_dicts
 
 
 class UtilsTest(unittest.TestCase):
